Extract file IO ops into separate file
