+++ CMakeLists.txt
@@ -7,6 +7,6 @@ add_executable(NameSort src/main.c
                         src/name_string.c
         src/string_vector.c
                         include/name_string.h
-                        include/string_vector.h)
+                        include/string_vector.h include/file_io.h src/file_io.c)
 
 target_include_directories(NameSort PRIVATE ${CMAKE_SOURCE_DIR}/include)
